Move systems into their own modules
